category add/update logic complete

+++ routes/categories.js
@@ -2,68 +2,80 @@ const express = require("express");
 const router = express.Router();
 const Category = require("../models/category");
 const SubCategory = require("../models/subCategory");
+const Product = require("../models/product");
 
 router.post("/", async (req, res) => {
 
-  const { mainCategory, subCategory } = req.body;
+  try{
+    const { mainCategory, subCategory } = req.body;
 
-  // check duplicate
-  let existingCategory = await Category.findOne({ name: mainCategory });
-  if (existingCategory) {
-    return res.status(400).send("Category already exists.");
-  }
-
-  let catCount = await getCatCount();
-  // If not, save the new category
-  let category = new Category({
-    name: mainCategory,
-    sku: catCount.toString().padStart(4,'0')
-  });
-
-  category = await category.save();
-
-  let subCatCount = await getSubCatCount();
-
-  const addedSubCat = await Promise.all(subCategory.map(async (subCatName) => {
-    let existingSubCat = await SubCategory.findOne({ name: subCatName });
-    if (existingSubCat) {
-      return;
+    // check duplicate
+    let existingCategory = await Category.findOne({ name: mainCategory });
+    if (existingCategory) {
+      return res.status(400).send("Category already exists.");
     }
-  
-    // create sub-category
-    let subCat = new SubCategory({
-      name: subCatName,
-      sku: subCatCount.toString().padStart(4, '0'),
-      parent_id: category._id
+
+    let catCount = await getCatCount();
+    // If not, save the new category
+    let category = new Category({
+      name: mainCategory,
+      sku: catCount.toString().padStart(4,'0')
     });
 
-    subCatCount++;
-  
-    subCat = await subCat.save();
-  
-  
-    return subCat._id;
-  }));
+    category = await category.save();
 
-  category = await Category.findByIdAndUpdate(
-    category._id, 
-    {
-      $set: {
-        subCategories: addedSubCat
+    let subCatCount = await getSubCatCount();
+
+    const addedSubCat = await Promise.all(subCategory.map(async (subCatName) => {
+
+      let existingSubCat = await SubCategory.findOne({ name: subCatName.name });
+      if (existingSubCat) {
+        return;
       }
-    },
-    {
-      new: true
-    }
-  );
+      
+      // create sub-category
+      let subCat = new SubCategory({
+        name: subCatName.name,
+        sku: subCatCount.toString().padStart(4, '0'),
+        parent_id: category._id
+      });
 
-  res.send(category);
+      subCatCount++;
+      
+      subCat = await subCat.save();
+      return subCat._id;
+    }));
+
+
+    category = await Category.findByIdAndUpdate(
+      category._id, 
+      {
+        $set: {
+          subCategories: addedSubCat
+        }
+      },
+      {
+        new: true
+      }
+    );
+
+    res.send(category);
+
+  }catch(ex){
+    console.log(ex);
+    res.status(500).send("Error");
+  }
 });
 
 
 router.get("/", async (req, res) => {
-  const categories = await Category.find().sort("mainCategory").populate("subCategories");
-  res.send(categories);
+  try{
+    const categories = await SubCategory.find().sort("name").populate("parent_id");
+    res.send(categories);
+  }catch(ex){
+    console.log(ex);
+    res.status(500).send("Error");
+  }
 });
 
 router.get("/:subcategory", async (req, res) => {
@@ -94,7 +106,7 @@ router.delete("/:id", async (req, res) => {
 
 router.get("/subcategory/:subcategory", async (req, res) => {
   const { subcategory } = req.body; // Access subcategory from req.query
-  console.log(subcategory);
+
   try {
     const category = await Category.findOne({ subCategory: subcategory });
     if (!category) {
@@ -112,7 +124,7 @@ router.get("/subcategory/:subcategory", async (req, res) => {
 router.put("/:category_id", async (req, res) => {
 
   try{
-    
+
     const { mainCategory, subCategory } = req.body;
     
     const category = await Category.findById(req.params.category_id);
@@ -122,55 +134,61 @@ router.put("/:category_id", async (req, res) => {
     }
 
     let subCatCount = await getSubCatCount();
-
+    let newAddedSubCats = [];
     for (const sub of subCategory) {
       if (!sub._id) {
         // Create new sub-category
         const newSubCategory = new SubCategory({
           name: sub.name,
-          category_id: category._id,
+          parent_id: category._id,
           sku: subCatCount.toString().padStart(4,"0")
         });
         subCatCount++;
         await newSubCategory.save();
-        category.subCategories.push(newSubCategory._id);
+        newAddedSubCats.push(newSubCategory._id);
       } else {
         // Update existing sub-category
         await SubCategory.findByIdAndUpdate(sub._id, { name: sub.name });
       }
     }
 
-    const mismatchedCats = subCats.filter(subCat => {
-      // Check if the subCat is not new and has a valid _id
-      if (subCat._id !== null) {
-        // Check if the subCat is not present in mainCats
-        return !mainCats.find(mainCat => mainCat._id === subCat._id);
-      }
-      return false; // Filter out new cats and cats without _id
+    const missingSubCategories = category.subCategories.filter(originalSubCat => {
+      return !subCategory.some(updatedSubCat => {
+        return updatedSubCat._id === originalSubCat.toString();
+      });
     });
-
-    // Remove deleted sub-categories not referred in products
-    mismatchedCats.map(async (subId) => {
-      const referredInProducts = await Product.findOne({ category: subId });
-      if (!referredInProducts) {
-        category.subCategories.filter(subCat => subCat._id !== subId);
-        await SubCategory.findByIdAndDelete(subId);
-        return false;
-      }
-    });
-
-    console.log(category);
     
-    await category.save();
+    // Remove deleted sub-categories not referred in products
+    if(missingSubCategories.length > 0){
+      missingSubCategories.map(async (subId) => {
+        if(subId === undefined) return; 
+        const referredInProducts = await Product.findOne({ category: subId });
+        if (!referredInProducts) {
+          category.subCategories.filter((subCat)=>{
+            return subCat.toString() !== subId.toString()
+          });
+          await SubCategory.findByIdAndDelete(subId);
+        }
+      });
+    }
+    
+    // maybe name is changed
+    category.name = mainCategory;
+    if(newAddedSubCats.length > 0){
+      category.subCategories.push(newAddedSubCats);
+    }
+    
+    await Category.findByIdAndUpdate(req.params.category_id,{
+      $set: category
+    });
 
     res.status(200).json({ message: 'Category updated successfully' })
 
   }catch(ex){
-    console.error(error);
+    console.error(ex);
     res.status(500).json({ error: 'Internal server error' });
   }
 });
-
 
 // get sub categories by category ID
 router.get("/get_subs/:cat_id", async (req, res) => {
@@ -191,13 +209,23 @@ router.get("/get_subs/:cat_id", async (req, res) => {
 
 
 async function getSubCatCount(){
-  let count = await SubCategory.countDocuments({});
-  return count++;
+  let lastDoc = await SubCategory.find().limit(1).sort({$natural:-1})
+  if(lastDoc.length === 0){
+    return 0;
+  }
+  let count = parseInt(lastDoc[0].sku);
+  count++;
+  return count;
 }
 
 async function getCatCount(){
-  let count = await Category.countDocuments({});
-  return count++;
+  let lastDoc = await Category.find().limit(1).sort({$natural:-1})
+  if(lastDoc.length === 0){
+    return 0;
+  }
+  let count = parseInt(lastDoc[0].sku);
+  count++;
+  return count;
 }
